Styling and JSON formatting and language input issue
Alot of RND has been done on this task, the input type is dependable on the end-user as he has to change the language settings of the keyboard from his OS (Windows/Mac), then he can type in whatever language he wants.

Huzeifa took part in its RND as well, that's how we came to the conclusion.